(feat) change router structure
