+++ src/app/core/core.module.ts
@@ -1,11 +1,11 @@
 import {NgModule} from '@angular/core';
 import {CommonModule} from "@angular/common";
-import {SobvCoreRoutingModule} from "./core-routing.module";
 import {SobvHeaderComponent} from "./components/sobv-header/sobv-header.component";
 import {SobvPageNotFoundComponent} from "./components/sobv-page-not-found/sobv-page-not-found.component";
 import {SobvDashboardComponent} from "./components/sobv-dashboard/sobv-dashboard.component";
 import {NgbTooltipModule} from "@ng-bootstrap/ng-bootstrap";
-import { SobvSubHeaderComponent } from './components/sobv-sub-header/sobv-sub-header.component';
+import {SobvSubHeaderComponent} from './components/sobv-sub-header/sobv-sub-header.component';
+import {RouterModule} from "@angular/router";
 
 @NgModule({
   declarations: [
@@ -16,7 +16,7 @@ import { SobvSubHeaderComponent } from './components/sobv-sub-header/sobv-sub-he
   ],
   imports: [
     CommonModule,
-    SobvCoreRoutingModule,
+    RouterModule,
     NgbTooltipModule
   ],
   exports: [
