+++ src/app/polls/polls.module.ts
@@ -1,6 +1,5 @@
 import {NgModule} from '@angular/core';
 import {CommonModule} from '@angular/common';
-import {SobvPollsRoutingModule} from "./polls-routing.module";
 import {SobvPollModalComponent} from './components/sobv-poll-modal/sobv-poll-modal.component';
 import { SobvPollQuestionsComponent } from './components/sobv-poll-questions/sobv-poll-questions.component';
 import { SobvPollChoicesComponent } from './components/sobv-poll-choices/sobv-poll-choices.component';
@@ -15,7 +14,6 @@ import { CoreModule } from '../core/core.module';
     ],
   imports: [
     CommonModule,
-    SobvPollsRoutingModule,
     CoreModule,
   ]
 })
