+++ src/app/profile-serviceman/profile-serviceman.module.ts
@@ -1,6 +1,5 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import {SobvProfileServicemanRoutingModule} from "./profile-serviceman-routing.module";
 import {SobvProfileServicemanComponent} from "./components/sobv-profile-serviceman/sobv-profile-serviceman.component";
 import {CoreModule} from "../core/core.module";
 import {NgbDropdownModule} from "@ng-bootstrap/ng-bootstrap";
@@ -11,7 +10,6 @@ import {NgbDropdownModule} from "@ng-bootstrap/ng-bootstrap";
   ],
   imports: [
     CommonModule,
-    SobvProfileServicemanRoutingModule,
     CoreModule,
     NgbDropdownModule
   ],
