+++ src/app/sobv-routing.module.ts
@@ -1,10 +1,22 @@
-import { NgModule } from '@angular/core';
-import { RouterModule, Routes } from '@angular/router';
+import {NgModule} from '@angular/core';
+import {RouterModule, Routes} from '@angular/router';
+import {SobvDashboardComponent} from "./core/components/sobv-dashboard/sobv-dashboard.component";
+import {SobvPageNotFoundComponent} from "./core/components/sobv-page-not-found/sobv-page-not-found.component";
+import {SobvPollModalComponent} from "./polls/components/sobv-poll-modal/sobv-poll-modal.component";
+import {SobvProfileServicemanComponent} from "./profile-serviceman/components/sobv-profile-serviceman/sobv-profile-serviceman.component";
 
-const routes: Routes = [];
+const routes: Routes = [
+  {path: 'dashboard', component: SobvDashboardComponent},
+  {path: '', redirectTo: '/dashboard', pathMatch: 'full'},
+  {path: 'polls/category/:categoryId/poll/:pollId', component: SobvPollModalComponent},
+  {path: 'profile/serviceman/:id', component: SobvProfileServicemanComponent},
+  {path: 'error/:code', component: SobvPageNotFoundComponent},
+  {path: '**', component: SobvPageNotFoundComponent}
+];
 
 @NgModule({
   imports: [RouterModule.forRoot(routes)],
   exports: [RouterModule]
 })
-export class SobvRoutingModule { }
+export class SobvRoutingModule {
+}
